research/1.131-project-management/02-implementations/vikunja-populate-script/src/requirements.txt
pyyaml>=6.0
pytest>=7.0
python-dotenv>=1.0
requests>=2.31.0
